backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiomysql==0.1.1
    # via -r requirements.in
aioredis==2.0.1
    # via -r requirements.in
anyio==3.7.0
    # via starlette
async-timeout==4.0.2
    # via aioredis
click==8.1.3
    # via uvicorn
exceptiongroup==1.1.1
    # via anyio
fastapi==0.96.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
pydantic==1.10.9
    # via fastapi
pymysql==1.0.3
    # via aiomysql
python-dotenv==1.0.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.6.3
    # via
    #   aioredis
    #   pydantic
ujson==5.7.0
    # via -r requirements.in
uvicorn==0.22.0
    # via -r requirements.in
